audit feedback implementation

+++ src/controllers/partyController.js
@@ -1,5 +1,6 @@
 import db from '../db';
 import { partyValidator } from '../middleware/schemaValidators';
+import auth from '../middleware/auth';
 
 class PartyController {
   /**
@@ -15,41 +16,47 @@ class PartyController {
    * @memberof PartyController
    */
   static async createParty(req, res) {
-    const { name, hqAddress, logoUrl } = req.body;
-    const createQuery = `INSERT INTO 
-      parties(name, hqAddress, logoUrl)
-      VALUES ($1, $2, $3)
-      RETURNING *`;
-    const values = [name.trim(), hqAddress.trim(), logoUrl.trim()];
+    // Gets the passed in the token generated on authentication
+    const user = auth.tokenBearer(req);
+    if (user.isAdmin === true) {
+      const { name, hqAddress, logoUrl } = req.body;
+      const createQuery = `INSERT INTO parties(name, hqAddress, logoUrl)
+      VALUES ($1, $2, $3) RETURNING *`;
+      const values = [name.trim(), hqAddress.trim(), logoUrl.trim()];
 
-    // validate the request from consumer.
-    const validate = partyValidator(req.body);
-    if (validate.error) {
-      const errorMessage = validate.error.details.map(m => m.message.replace(/[^a-zA-Z0-9 ]/g, ''),);
-      return res.status(422).json({
-        status: 422,
-        error: errorMessage,
-      });
-    }
-
-    try {
-      const { rows } = await db.query(createQuery, values);
-      return res.status(201).send({
-        status: 201,
-        data: [rows[0]],
-      });
-    } catch (error) {
-      if (error.routine === '_bt_check_unique') {
-        return res.status(400).send({
-          status: 400,
-          error: `Party with '${name.trim()}' already exists`,
+      // validate the request from consumer.
+      const validate = partyValidator(req.body);
+      if (validate.error) {
+        const errorMessage = validate.error.details.map(m => m.message.replace(/[^a-zA-Z0-9 ]/g, ''));
+        return res.status(422).json({
+          status: 422,
+          error: errorMessage,
         });
       }
-      return res.status(500).send({
-        status: 500,
-        error: 'Unable to Create Party!! Server Error, Try Again',
-      });
+
+      try {
+        const { rows } = await db.query(createQuery, values);
+        return res.status(201).send({
+          status: 201,
+          data: [rows[0]],
+        });
+      } catch (error) {
+        if (error.routine === '_bt_check_unique') {
+          return res.status(400).send({
+            status: 400,
+            error: `Party with '${name.trim()}' already exists`,
+          });
+        }
+        return res.status(500).send({
+          status: 500,
+          error: 'Unable to Create Party!! Server Error, Try Again',
+        });
+      }
     }
+    return res.status(401).json({
+      status: 401,
+      error: 'Unauthorized',
+    });
   }
 
   /**
@@ -129,46 +136,54 @@ class PartyController {
    * @memberof PartyController
    */
   static async editParty(req, res) {
-    const getPartyQuery = 'SELECT * FROM  parties WHERE party_id = $1';
-    const editQuery = `UPDATE parties 
+    // Gets the passed in the token generated on authentication
+    const user = auth.tokenBearer(req);
+    if (user.isAdmin === true) {
+      const getPartyQuery = 'SELECT * FROM  parties WHERE party_id = $1';
+      const editQuery = `UPDATE parties 
       SET name=$1, hqAddress=$2, logoURL=$3
       WHERE party_id=$4 RETURNING *`;
 
-    // validates the data sent by consumer
-    const validate = partyValidator(req.body);
-    if (validate.error) {
-      const errorMessage = validate.error.details.map(m => m.message.replace(/[^a-zA-Z0-9 ]/g, ''),);
-      return res.status(422).json({
-        status: 422,
-        error: errorMessage,
-      });
-    }
-    try {
-      const { rows } = await db.query(getPartyQuery, [req.params.id]);
-      if (!rows[0]) {
-        return res.status(404).send({
-          status: 404,
-          data: 'Party does not exist',
+      // validates the data sent by consumer
+      const validate = partyValidator(req.body);
+      if (validate.error) {
+        const errorMessage = validate.error.details.map(m => m.message.replace(/[^a-zA-Z0-9 ]/g, ''));
+        return res.status(422).json({
+          status: 422,
+          error: errorMessage,
         });
       }
-      const { name, hqAddress, logoUrl } = req.body;
-      const values = [
-        name.trim(),
-        hqAddress.trim(),
-        logoUrl.trim(),
-        req.params.id,
-      ];
-      const response = await db.query(editQuery, values);
-      return res.status(200).send({
-        status: 200,
-        data: [response.rows[0]],
-      });
-    } catch (error) {
-      return res.status(400).send({
-        status: 400,
-        error: 'Party Edit Not Saved!! Server Error, Please Try Again',
-      });
+      try {
+        const { rows } = await db.query(getPartyQuery, [req.params.id]);
+        if (!rows[0]) {
+          return res.status(404).send({
+            status: 404,
+            data: 'Party does not exist',
+          });
+        }
+        const { name, hqAddress, logoUrl } = req.body;
+        const values = [
+          name.trim(),
+          hqAddress.trim(),
+          logoUrl.trim(),
+          req.params.id,
+        ];
+        const response = await db.query(editQuery, values);
+        return res.status(200).send({
+          status: 200,
+          data: [response.rows[0]],
+        });
+      } catch (error) {
+        return res.status(400).send({
+          status: 400,
+          error: 'Party Edit Not Saved!! Server Error, Please Try Again',
+        });
+      }
     }
+    return res.status(401).json({
+      status: 401,
+      error: 'Unauthorized',
+    });
   }
 
   /**
@@ -184,25 +199,33 @@ class PartyController {
    * @memberof PartyController
    */
   static async deleteParty(req, res) {
-    const deleteQuery = 'DELETE FROM parties WHERE party_id=$1 RETURNING *';
-    try {
-      const { rows } = await db.query(deleteQuery, [req.params.id]);
-      if (!rows[0]) {
-        return res.status(404).send({
-          status: 404,
-          data: 'Party does not exitst',
+    // Gets the passed in the token generated on authentication
+    const user = auth.tokenBearer(req);
+    if (user.isAdmin === true) {
+      const deleteQuery = 'DELETE FROM parties WHERE party_id=$1 RETURNING *';
+      try {
+        const { rows } = await db.query(deleteQuery, [req.params.id]);
+        if (!rows[0]) {
+          return res.status(404).send({
+            status: 404,
+            data: 'Party does not exitst',
+          });
+        }
+        return res.status(200).send({
+          status: 204,
+          data: 'Item deleted',
+        });
+      } catch (error) {
+        return res.status(500).send({
+          status: 500,
+          error: 'Party delete not completed!! Server Error, Please Try Again',
         });
       }
-      return res.status(200).send({
-        status: 204,
-        data: 'Item deleted',
-      });
-    } catch (error) {
-      return res.status(500).send({
-        status: 500,
-        error: 'Party delete not completed!! Server Error, Please Try Again',
-      });
     }
+    return res.status(401).json({
+      status: 401,
+      error: 'Unauthorized',
+    });
   }
 }
